chore(camera): move LocalCameraRepository implementation to a new file
